+++ packages/in_app_purchase/in_app_purchase_storekit/lib/src/store_kit_wrappers/sk_request_maker.dart
@@ -5,11 +5,12 @@
 import 'dart:async';
 
 import 'package:flutter/services.dart';
-import 'package:flutter/widgets.dart';
 
 import '../channel.dart';
 import '../messages.g.dart';
 import 'sk_product_wrapper.dart';
+
+InAppPurchaseAPI _hostApi = InAppPurchaseAPI();
 
 /// A request maker that handles all the requests made by SKRequest subclasses.
 ///
@@ -17,14 +18,6 @@ import 'sk_product_wrapper.dart';
 /// we consolidated all the `SKRequest` subclasses into this class to make requests in a more straightforward way.
 /// The request maker will create a SKRequest object, immediately starting it, and completing the future successfully or throw an exception depending on what happened to the request.
 class SKRequestMaker {
-  InAppPurchaseAPI _hostApi = InAppPurchaseAPI();
-
-  /// Set up pigeon API.
-  @visibleForTesting
-  void setInAppPurchaseHostApi(InAppPurchaseAPI api) {
-    _hostApi = api;
-  }
-
   /// Fetches product information for a list of given product identifiers.
   ///
   /// The `productIdentifiers` should contain legitimate product identifiers that you declared for the products in the iTunes Connect. Invalid identifiers