+++ packages/in_app_purchase/in_app_purchase_storekit/pigeons/messages.dart
@@ -256,6 +256,4 @@ abstract class InAppPurchaseAPI {
   void restoreTransactions(String? applicationUserName);
 
   void presentCodeRedemptionSheet();
-  //
-  // void retrieveReceiptData();
 }