+++ packages/in_app_purchase/in_app_purchase_storekit/test/fakes/fake_storekit_platform.dart
@@ -124,42 +124,6 @@ class FakeStoreKitPlatform implements TestInAppPurchaseApi {
 
   Future<dynamic> onMethodCall(MethodCall call) {
     switch (call.method) {
-      case '-[InAppPurchasePlugin startProductRequest:result:]':
-        if (queryProductException != null) {
-          throw queryProductException!;
-        }
-        final List<String> productIDS =
-            List.castFrom<dynamic, String>(call.arguments as List<dynamic>);
-        final List<String> invalidFound = <String>[];
-        final List<SKProductWrapper> products = <SKProductWrapper>[];
-        for (final String productID in productIDS) {
-          if (!validProductIDs.contains(productID)) {
-            invalidFound.add(productID);
-          } else {
-            products.add(validProducts[productID]!);
-          }
-        }
-        final SkProductResponseWrapper response = SkProductResponseWrapper(
-            products: products, invalidProductIdentifiers: invalidFound);
-        return Future<Map<String, dynamic>>.value(
-            buildProductResponseMap(response));
-      case '-[InAppPurchasePlugin restoreTransactions:result:]':
-        if (restoreException != null) {
-          throw restoreException!;
-        }
-        if (testRestoredError != null) {
-          InAppPurchaseStoreKitPlatform.observer
-              .restoreCompletedTransactionsFailed(error: testRestoredError!);
-          return Future<void>.sync(() {});
-        }
-        if (!testRestoredTransactionsNull) {
-          InAppPurchaseStoreKitPlatform.observer
-              .updatedTransactions(transactions: transactionList);
-        }
-        InAppPurchaseStoreKitPlatform.observer
-            .paymentQueueRestoreCompletedTransactionsFinished();
-
-        return Future<void>.sync(() {});
       case '-[InAppPurchasePlugin retrieveReceiptData:result:]':
         if (receiptData != null) {
           return Future<String>.value(receiptData!);
@@ -169,53 +133,6 @@ class FakeStoreKitPlatform implements TestInAppPurchaseApi {
       case '-[InAppPurchasePlugin refreshReceipt:result:]':
         receiptData = 'refreshed receipt data';
         return Future<void>.sync(() {});
-      case '-[InAppPurchasePlugin addPayment:result:]':
-        final Map<String, Object?> arguments = _getArgumentDictionary(call);
-        final String id = arguments['productIdentifier']! as String;
-        final int quantity = arguments['quantity']! as int;
-
-        // Keep the received paymentDiscount parameter when testing payment with discount.
-        if (arguments['applicationUsername']! == 'userWithDiscount') {
-          final Map<dynamic, dynamic>? discountArgument =
-              arguments['paymentDiscount'] as Map<dynamic, dynamic>?;
-          if (discountArgument != null) {
-            discountReceived = discountArgument.cast<String, dynamic>();
-          } else {
-            discountReceived = <String, dynamic>{};
-          }
-        }
-
-        final SKPaymentTransactionWrapper transaction =
-            createPendingTransaction(id, quantity: quantity);
-        transactionList.add(transaction);
-        InAppPurchaseStoreKitPlatform.observer.updatedTransactions(
-            transactions: <SKPaymentTransactionWrapper>[transaction]);
-        sleep(const Duration(milliseconds: 30));
-        if (testTransactionFail) {
-          final SKPaymentTransactionWrapper transactionFailed =
-              createFailedTransaction(id, quantity: quantity);
-          InAppPurchaseStoreKitPlatform.observer.updatedTransactions(
-              transactions: <SKPaymentTransactionWrapper>[transactionFailed]);
-        } else if (testTransactionCancel > 0) {
-          final SKPaymentTransactionWrapper transactionCanceled =
-              createCanceledTransaction(id, testTransactionCancel,
-                  quantity: quantity);
-          InAppPurchaseStoreKitPlatform.observer.updatedTransactions(
-              transactions: <SKPaymentTransactionWrapper>[transactionCanceled]);
-        } else {
-          final SKPaymentTransactionWrapper transactionFinished =
-              createPurchasedTransaction(
-                  id, transaction.transactionIdentifier ?? '',
-                  quantity: quantity);
-          InAppPurchaseStoreKitPlatform.observer.updatedTransactions(
-              transactions: <SKPaymentTransactionWrapper>[transactionFinished]);
-        }
-      case '-[InAppPurchasePlugin finishTransaction:result:]':
-        final Map<String, Object?> arguments = _getArgumentDictionary(call);
-        finishedTransactions.add(createPurchasedTransaction(
-            arguments['productIdentifier']! as String,
-            arguments['transactionIdentifier']! as String,
-            quantity: transactionList.first.payment.quantity));
       case '-[SKPaymentQueue startObservingTransactionQueue]':
         queueIsActive = true;
       case '-[SKPaymentQueue stopObservingTransactionQueue]':
@@ -291,24 +208,54 @@ class FakeStoreKitPlatform implements TestInAppPurchaseApi {
 
   @override
   void finishTransaction(Map<String?, String?> finishMap) {
-    // TODO: implement finishTransaction
+    finishedTransactions.add(createPurchasedTransaction(
+        finishMap['productIdentifier']!, finishMap['transactionIdentifier']!,
+        quantity: transactionList.first.payment.quantity));
   }
 
   @override
   void presentCodeRedemptionSheet() {
-    // TODO: implement presentCodeRedemptionSheet
+    return;
   }
 
   @override
   void restoreTransactions(String? applicationUserName) {
-    // TODO: implement restoreTransactions
+    if (restoreException != null) {
+      throw restoreException!;
+    }
+    if (testRestoredError != null) {
+      InAppPurchaseStoreKitPlatform.observer
+          .restoreCompletedTransactionsFailed(error: testRestoredError!);
+      return;
+    }
+    if (!testRestoredTransactionsNull) {
+      InAppPurchaseStoreKitPlatform.observer
+          .updatedTransactions(transactions: transactionList);
+    }
+    InAppPurchaseStoreKitPlatform.observer
+        .paymentQueueRestoreCompletedTransactionsFinished();
   }
 
   @override
   Future<SKProductsResponseMessage> startProductRequest(
       List<String?> productIdentifiers) {
-    print("hello 2");
+    if (queryProductException != null) {
+      throw queryProductException!;
+    }
+    final List<String?> productIDS = productIdentifiers;
+    final List<String> invalidFound = <String>[];
+    final List<SKProductWrapper> products = <SKProductWrapper>[];
+    for (final String? productID in productIDS) {
+      if (!validProductIDs.contains(productID)) {
+        invalidFound.add(productID!);
+      } else {
+        products.add(validProducts[productID]!);
+      }
+    }
+    final SkProductResponseWrapper response = SkProductResponseWrapper(
+        products: products, invalidProductIdentifiers: invalidFound);
 
-    throw UnimplementedError();
+    return Future<SKProductsResponseMessage>.value(
+        SkProductResponseWrapper.convertToPigeon(response));
   }
 }