+++ packages/in_app_purchase/in_app_purchase_storekit/test/store_kit_wrappers/sk_test_stub_objects.dart
@@ -153,8 +153,6 @@ final SKProductsResponseMessage dummyProductResponseMessage =
   invalidProductIdentifiers: const <String>['123'],
 );
 
-class SkProductsResponseMessage {}
-
 Map<String, dynamic> buildLocaleMap(SKPriceLocaleWrapper local) {
   return <String, dynamic>{
     'currencySymbol': local.currencySymbol,